+++ contracts/LiveLibsUtils.sol
@@ -1,7 +1,5 @@
-// Gonna need to link this
-
 library LiveLibsUtils {
-    function toVersionNum(uint8 major, uint8 minor, uint8 patch) constant returns (uint) {
+    function toVersionNum(uint major, uint minor, uint patch) constant returns (uint) {
         return 1000000*major + 1000*minor + patch;
     }
 }
